requirements.txt
boto3==1.18.60
Pillow==8.4.0